+++ src/main/java/com/egyptFlightReservation/Model/Payment/PaymentMethod.java
@@ -1,14 +1,14 @@
 package com.egyptFlightReservation.Model.Payment;
 
 public abstract class PaymentMethod implements PaymentProcessor {
-    protected int curCash;
+    protected double curCash;
     public static final int DEFAULT_CASH = 50000;
 
-    public PaymentMethod(int curCash) {
+    public PaymentMethod(double curCash) {
         this.curCash = curCash;
     }
 
-    public int getCurCash() {
+    public double getCurCash() {
         return curCash;
     }
