+++ src/main/java/com/egyptFlightReservation/Model/Payment/PaymentProcessor.java
@@ -1,5 +1,5 @@
 package com.egyptFlightReservation.Model.Payment;
 
 public interface PaymentProcessor {
-    boolean paymentProcess(int amount);
+    boolean paymentProcess(double amount);
 }
